SocialServices-Refactoring: Moved services implementation to impl package.
