+++ src/main/java/com/umssonline/social/services/impl/CommentServiceImpl.java
@@ -1,8 +1,9 @@
-package com.umssonline.social.services;
+package com.umssonline.social.services.impl;
 
 import com.umssonline.social.models.entity.Comment;
 import com.umssonline.social.models.entity.Participant;
 import com.umssonline.social.repositories.api.ExtendedCommentDao;
+import com.umssonline.social.services.CrudSocialService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
