+++ src/main/java/com/umssonline/social/services/impl/ParticipantServiceImpl.java
@@ -1,7 +1,8 @@
-package com.umssonline.social.services;
+package com.umssonline.social.services.impl;
 
 import com.umssonline.social.models.entity.Participant;
 import com.umssonline.social.repositories.api.ExtendedParticipantDao;
+import com.umssonline.social.services.CrudSocialService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
